Update and expand the README
I clarified and expanded the documentation in the project's README file.
As the first line of documentation for the project, it's important to
have a clear and complete README file.

Amoung various minor expansions and clarifications, I added
documentation for the hidden feature of multi-level wildcards (wildcards
that match over multiple levels of namespaces, defined with a
double asterisk). This is a feature that as best I can tell isn't
documented anywhere (at least not anywhere obvious), and the lack of
accessible documentation of the feature seems to cause some grief. (For
example, see issue #154. It's not that the requested feature doesn't
exist, it's that the documentation is strangely silent about this
feature.)